% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{clipper_BC}
\alias{clipper_BC}
\title{FDR control method based on clipper.}
\usage{
clipper_BC(contrastScore, FDR)
}
\arguments{
\item{contrastScore}{Contrast Score or fliter statistics;}

\item{FDR}{Scalar or vector, target control level.}
}
\value{
FDR: If is vector, output FDR value of result;
 thre: Control thresholds for contrast scores;
 q: The converted values of the scores were compared;
 discovery: Results of variable selection at the FDR level.
}
\description{
FDR control method based on clipper.
}
