% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{contrast_score_computation}
\alias{contrast_score_computation}
\title{Calculation of test statistics based on DE (wilcoxon, ks.test).}
\usage{
contrast_score_computation(W, result3, class0, test1)
}
\arguments{
\item{W}{Count matrix;}

\item{result3}{knockoff generation.}

\item{class0}{Response variable, binary variable or others.}

\item{test1}{Selection of test statistics, c("wilcox.test", "ks.test")}
}
\value{
contrast_score: contrast score.
}
\description{
Calculation of test statistics based on DE (wilcoxon, ks.test).
}
