% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{FDR_Power}
\alias{FDR_Power}
\title{Calculation of FDR and Power}
\usage{
FDR_Power(a_1, b_1 = NULL)
}
\arguments{
\item{a_1}{Results of selected variables;}

\item{b_1}{True significant variable.}
}
\value{
FDR and Power.
}
\description{
Calculation of FDR and Power
}
