% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{fit_Gaussian_copula}
\alias{fit_Gaussian_copula}
\title{Fit a Gaussian copula model for taxa counts}
\usage{
fit_Gaussian_copula(
  x,
  marginal = c("auto_choose", "zinb", "nb", "poisson"),
  jitter = TRUE,
  zp_cutoff = 0.8,
  min_nonzero_num = 2
)
}
\arguments{
\item{x}{Count matrix with \code{p × n}.}

\item{marginal}{The types of marginal distribution.
Default value is 'auto_choose' which chooses between ZINB, NB, ZIP
and Poisson by a likelihood ratio test (lrt) and whether there is
underdispersion.
'zinb' will fit the ZINB model. If there is underdispersion, it
will choose between ZIP and Poisson by a lrt. Otherwise, it will try to
fit the ZINB model. If in this case, there is no zero at all or an error
occurs, it will fit an NB model instead.
'nb' fits the NB model that chooses between NB and Poisson depending
on whether there is underdispersion.
'poisson' simply fits the Poisson model.}

\item{jitter}{Logical, whether a random projection should be performed in the
distributional transform.}

\item{zp_cutoff}{Proportion threshold of zeros; taxa with zero proportion
below this cutoff are included in the joint copula model.}

\item{min_nonzero_num}{Minimum number of non-zero counts required for a taxon
to fit a marginal model.}
}
\value{
A list describing three taxon groups and the fitted copula model:
\describe{
  \item{cov_mat}{Correlation matrix of the Gaussian copula (group 1 taxa).}
  \item{marginal_param1}{Marginal parameters for group 1 taxa.}
  \item{marginal_param2}{Marginal parameters for group 2 taxa.}
  \item{taxa_sel1}{Indices of taxa included in the copula model.}
  \item{taxa_sel2}{Indices of taxa with only marginal models.}
  \item{taxa_sel3}{Indices of remaining taxa.}
  \item{zp_cutoff}{Input zero-proportion cutoff.}
  \item{min_nonzero_num}{Input minimum non-zero requirement.}
  \item{sim_method}{Character string \code{"copula"}.}
}
}
\description{
Fit a Gaussian copula model for taxa counts
}
