% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{fit_model_scDesign2}
\alias{fit_model_scDesign2}
\title{Fit models for microbiome count matrix}
\usage{
fit_model_scDesign2(
  data_mat,
  sam_type_sel,
  sim_method = c("copula", "ind"),
  marginal = c("auto_choose", "zinb", "nb", "poisson"),
  jitter = TRUE,
  zp_cutoff = 0.8,
  min_nonzero_num = 2,
  ncores = 1
)
}
\arguments{
\item{data_mat}{A p × n count matrix (p taxa, n samples).
Column names should indicate sample-type (e.g., condition or group)
and must match \code{sam_type_sel}.}

\item{sam_type_sel}{Character vector of selected sample-types for model fitting.}

\item{sim_method}{Type of model to fit: 'copula' (default) or 'ind' (independent).}

\item{marginal}{The types of marginal distribution.
Default value is 'auto_choose' which chooses between ZINB, NB, ZIP
and Poisson by a likelihood ratio test (lrt) and whether there is
underdispersion.
'zinb' will fit the ZINB model. If there is underdispersion, it
will choose between ZIP and Poisson by a lrt. Otherwise, it will try to
fit the ZINB model. If in this case, there is no zero at all or an error
occurs, it will fit an NB model instead.
'nb' fits the NB model that chooses between NB and Poisson depending
on whether there is underdispersion.
'poisson' simply fits the Poisson model.}

\item{jitter}{Logical, whether a random projection should be performed in the
distributional transform.}

\item{zp_cutoff}{Proportion threshold of zeros; taxa with zero proportion
below this cutoff are included in the joint copula model.}

\item{min_nonzero_num}{Minimum number of non-zero counts required for a taxon
to fit a marginal model.}

\item{ncores}{Number of parallel cores.}
}
\value{
A list (length = number of sample types), where each element contains the fitted
        model for that sample type.
}
\description{
Fit models for microbiome count matrix
}
