% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{fit_wo_copula}
\alias{fit_wo_copula}
\title{Fit an independent marginal model for taxa counts}
\usage{
fit_wo_copula(
  x,
  marginal = c("auto_choose", "zinb", "nb", "poisson"),
  jitter = TRUE,
  min_nonzero_num = 2
)
}
\arguments{
\item{x}{Count matrix with \code{p × n}.}

\item{marginal}{The types of marginal distribution.
Default value is 'auto_choose' which chooses between ZINB, NB, ZIP
and Poisson by a likelihood ratio test (lrt) and whether there is
underdispersion.
'zinb' will fit the ZINB model. If there is underdispersion, it
will choose between ZIP and Poisson by a lrt. Otherwise, it will try to
fit the ZINB model. If in this case, there is no zero at all or an error
occurs, it will fit an NB model instead.
'nb' fits the NB model that chooses between NB and Poisson depending
on whether there is underdispersion.
'poisson' simply fits the Poisson model.}

\item{jitter}{Logical, whether a random projection should be performed in the
distributional transform.}

\item{min_nonzero_num}{Minimum number of non-zero counts required for a taxon
to fit a marginal model.}
}
\value{
A list containing the fitted independent marginal model:
\describe{
  \item{marginal_param1}{Marginal parameters for taxa with at least
    \code{min_nonzero_num} non-zero counts.}
  \item{taxa_sel1}{Indices of taxa with sufficient non-zero counts.}
  \item{taxa_sel2}{Indices of taxa excluded from modeling.}
  \item{min_nonzero_num}{Input threshold for minimum non-zero counts.}
  \item{sim_method}{The string \code{"ind"} indicating an independent model.}
}
}
\description{
This function only fits the marginal distribution for each taxon.
}
