% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{glm.nb}
\alias{glm.nb}
\title{glm.nb function: Estimation of the parameter (theta) for a Negative Binomial generalized linear model.}
\usage{
glm.nb(
  formula,
  data,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  control = glm.control(...),
  method = "glm.fit",
  model = TRUE,
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  ...,
  init.theta,
  link = log
)
}
\description{
glm.nb function: Estimation of the parameter (theta) for a Negative Binomial generalized linear model.
}
\keyword{internal}
