% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{knockoff.filter}
\alias{knockoff.filter}
\title{knockoff.filter}
\usage{
knockoff.filter(
  X,
  y,
  knockoffs = create.second_order,
  statistic = stat.glmnet_coefdiff,
  fdr = 0.1,
  offset = 1
)
}
\description{
knockoff.filter
}
