% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{mvrnorm}
\alias{mvrnorm}
\title{mvrnorm function: Produces samples from the specified multivariate normal distribution}
\usage{
mvrnorm(n = 1, mu, Sigma, tol = 1e-06, empirical = FALSE, EISPACK = FALSE)
}
\description{
mvrnorm function: Produces samples from the specified multivariate normal distribution
}
\keyword{internal}
