% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{negative.binomial}
\alias{negative.binomial}
\title{negative.binomial function: Fit a Negative Binomial generalized linear model with (theta).}
\usage{
negative.binomial(theta = stop("'theta' must be specified"), link = "log")
}
\description{
negative.binomial function: Fit a Negative Binomial generalized linear model with (theta).
}
\keyword{internal}
