% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_merge_all_dyads.R
\name{op_merge_all_dyads}
\alias{op_merge_all_dyads}
\title{Merge all combined dyad CSV files into one}
\usage{
op_merge_all_dyads(input_base_path, output_file)
}
\arguments{
\item{input_base_path}{Character. The base path containing the merged dyad CSV files.}

\item{output_file}{Character. The path to the output CSV file.}
}
\value{
None. The function is called for its side effects.
}
\description{
This function merges all combined dyad CSV files from the specified input base path into a single CSV file.
}
\examples{
\dontrun{
input_base_path <- "/path/to/OpenPose_Merged_CSV"
output_file <- "/path/to/All_Dyads_Merged.csv"
merge_all_dyads(input_base_path, output_file)
}
}
