% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_merge_each_dyad.R
\name{op_merge_each_dyad}
\alias{op_merge_each_dyad}
\title{Merge CSV files for each dyad}
\usage{
op_merge_each_dyad(input_base_path, output_base_path)
}
\arguments{
\item{input_base_path}{Character. The base path containing dyad directories with CSV files.}

\item{output_base_path}{Character. The base path where the merged CSV files will be saved.}
}
\value{
None. The function is called for its side effects.
}
\description{
This function merges all CSV files in each dyad directory within the specified input base path.
}
\examples{
\dontrun{
input_base_path <- "/path/to/OpenPose_CSV"
output_base_path <- "/path/to/OpenPose_Merged_CSV"
merge_dyad_csvs(input_base_path, output_base_path)
}
}
