% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{scDesign2_simulation}
\alias{scDesign2_simulation}
\title{knockoff generation based on scDesign2 methods.}
\usage{
scDesign2_simulation(W, class0)
}
\arguments{
\item{W}{Count matrix.}

\item{class0}{Response variable, binary variable.}
}
\value{
sim_count_copula: knockoff generation.
}
\description{
knockoff generation based on scDesign2 methods.
}
