% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{simulate_count_copula}
\alias{simulate_count_copula}
\title{Coupla-based Count Simulation}
\usage{
simulate_count_copula(copula_result, n = 100, marginal = c("nb", "Gamma"))
}
\arguments{
\item{copula_result}{Coupla model parameters.}

\item{n}{Number of simulated samples.}

\item{marginal}{Marginal distribution: \code{"nb"} or \code{"Gamma"}.}
}
\value{
Simulated count matrix \code{p × n}
\code{copula_result}
}
\description{
Coupla-based Count Simulation
}
