% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{simulate_count_ind}
\alias{simulate_count_ind}
\title{Simulate a count matrix for a single sam type based on a (w/o copula model)}
\usage{
simulate_count_ind(model_params, n = 100, marginal = c("nb", "Gamma"))
}
\arguments{
\item{model_params}{A list containing marginal parameters for simulation}

\item{n}{Number of simulated samples.}

\item{marginal}{Marginal distribution: \code{"nb"} or \code{"Gamma"}.}
}
\value{
A \code{p × n} simulated count matrix based on marginal NB or Gamma
models, where \code{p} is derived from \code{model_params}.
}
\description{
Simulate a count matrix for a single sam type based on a (w/o copula model)
}
