% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{simulate_count_scDesign2}
\alias{simulate_count_scDesign2}
\title{Simulate count matrix for experimental design}
\usage{
simulate_count_scDesign2(
  model_params,
  n_sam_new,
  sam_type_prop = 1,
  total_count_new = NULL,
  total_count_old = NULL,
  n_sam_old = NULL,
  sim_method = c("copula", "ind"),
  reseq_method = c("mean_scale", "multinomial"),
  sam_type = FALSE
)
}
\arguments{
\item{model_params}{A list of fitted microbiome models (one per group or condition),
each containing parameters from either a copula-based or independent
marginal model.}

\item{n_sam_new}{The total number of samples in the simulated count matrix.}

\item{sam_type_prop}{The sam type proportion in the simulated count matrix.}

\item{total_count_new}{Total number of reads in the simulated count matrix.}

\item{total_count_old}{The total number of reads or in the original count matrix.}

\item{n_sam_old}{The The total number of samples in the original count matrix.}

\item{sim_method}{Specification of the type of model for data simulation.
Default value is 'copula', which selects the copula model.
'ind' will select the (w/o copula) model.}

\item{reseq_method}{Strategy for adjusting sequencing depth:
\code{"mean_scale"} (default) rescales marginal means;
\code{"multinomial"} performs multinomial resampling to match
\code{total_count_new}.}

\item{sam_type}{Logical, whether sams for each group should be sampled from a
multinomial distribution or follows the exact same proportion as
specified in \code{sam_type_prop}.}
}
\value{
A matrix of shape p by n that contains the simulated count values. p is derived from
\code{model_params}.
}
\description{
Simulate count matrix for experimental design
}
