% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_others.R
\name{theta.ml}
\alias{theta.ml}
\title{theta.ml() function: Estimate (theta) of the Negative Binomial Distribution.}
\usage{
theta.ml(
  y,
  mu,
  n = sum(weights),
  weights,
  limit = 10,
  eps = .Machine$double.eps^0.25,
  trace = FALSE
)
}
\description{
theta.ml() function: Estimate (theta) of the Negative Binomial Distribution.
}
\keyword{internal}
